Rename dev-server-config -> dev-server
