+++ src/exports/index.js
@@ -5,7 +5,7 @@
 /** @typedef {{input?: InputOption, outDir?: ResolvedConfig['build']['outDir']}} Options */
 
 import { createLogger, mergeConfig } from 'vite';
-import { dev_server_config } from './plugins/dev-server-config.js';
+import { dev_server_config } from './plugins/dev-server.js/index.js';
 import { dev_server_manifest } from './plugins/dev-server-manifest.js';
 
 /**
