+++ src/exports/plugins/index.js
@@ -1,3 +1,3 @@
-export { dev_server_config } from './dev-server-config.js';
+export { dev_server_config } from './dev-server.js';
 export { dev_server_manifest } from './dev-server-manifest.js';
 export { wp_scripts } from './wp-scripts.js';
